+++ src/components/dashboard/Recruiterdashboard/DashboardNavbar.jsx
@@ -2,13 +2,13 @@ import { CgProfile } from 'react-icons/cg'
 import { MdDashboard, MdOutlinePostAdd } from 'react-icons/md'
 import Link from 'next/link'
 
-const Navbar = () => {
+const DashboardNavbar = () => {
   return (
-    <div className="w-[15%] shadow-lg rounded-3xl  px-4 py-6">
+    <div className="w-[25%] shadow-lg rounded-3xl px-4 py-6 hidden md:block">
       <ul className="flex flex-col  gap-4 items-start justify-center ">
         <li className="flex w-full items-center flex-1 border-b pb-2">
           <MdDashboard className="text-2xl inline-block mr-3" />
-          <span className="">Dash board</span>
+          <span className="font-bold">Dashboard</span>
         </li>
         <li className="w-full border-b pb-2">
           <Link
@@ -16,7 +16,7 @@ const Navbar = () => {
             className="flex items-center flex-1 "
           >
             <CgProfile className="text-2xl inline-block mr-3" />
-            <span className="">Profile</span>
+            <span className="font-semibold">Profile</span>
           </Link>
         </li>
         <li className="w-full border-b pb-2">
@@ -25,7 +25,7 @@ const Navbar = () => {
             className="flex items-center"
           >
             <MdOutlinePostAdd className="text-2xl inline-block mr-3" />
-            <span>Post a job</span>
+            <span className="font-semibold">Post a job</span>
           </Link>
         </li>
       </ul>
@@ -33,4 +33,4 @@ const Navbar = () => {
   )
 }
 
-export default Navbar
+export default DashboardNavbar
